+++ src/components/ImageExportModal.tsx
@@ -6,7 +6,8 @@ import { generateImage } from '../utils/imageExport';
 import { useTournamentStore, useCurrentGroup } from '../store/useTournamentStore';
 import { useEscapeClose } from '../hooks/useEscapeClose';
 import type { TournamentGroup } from '../types';
-import { useLanguagePreference, formatText } from '../i18n';
+import { useLanguagePreference } from '../i18nContext';
+import { formatText } from '../i18nData';
 
 type ExportType = 'ranking' | 'match';